+++ Splitgood/views.py
@@ -1,12 +1,8 @@
-from io import BytesIO
-
 from django.contrib.auth.models import User
 from django.core.paginator import Paginator
 from django.shortcuts import render, redirect
 from django.contrib.auth import logout, authenticate, login
 from django.views import View
-from django.views.decorators.csrf import csrf_protect
-
 from group.models import Group
 from userprofile.models import Profile
 
@@ -84,5 +80,8 @@ class InvitedRegisterView(View):
         login(request, user)
         group.users.add(user)
         groups = Group.objects.filter(users=request.user.id)
-        context = {'groups': groups, 'group_name': group_name}
+        paginator = Paginator(groups, 6)  # Create a Paginator instance with 6 items per page
+        page_number = request.GET.get('page', 1)  # Get the current page number from the request's query parameters
+        page_obj = paginator.get_page(page_number)
+        context = {'page_obj': page_obj}
         return render(request, "dashboard.html", context)
