+++ activites/urls.py
@@ -1,7 +1,8 @@
+from django.contrib.auth.decorators import login_required
 from django.urls import path
 
 from activites.views import show_activity
 
 urlpatterns = [
-    path("show_activities/", show_activity),
+    path("show_activities/", login_required(show_activity, login_url='/')),
 ]
