+++ static/css/list_expense.css
@@ -1,91 +1,104 @@
 *,
 *::before
 *::after {
-  margin: 0;
-  padding: 0;
-  box-sizing: inherit;
+    margin: 0;
+    padding: 0;
+    box-sizing: inherit;
 }
 
 body {
-  background-color: #f5f5f5;
+    background-color: #f5f5f5;
 }
+
 @media (max-width: 60em) {
-  body {
-    padding: 3rem;
-  }
+    body {
+        padding: 3rem;
+    }
 }
 
 .grid {
-  display: -webkit-box;
-  width: 85rem;
-  grid-gap: 6rem;
-  grid-template-columns: repeat(auto-fit, minmax(30rem, 1fr));
-  align-items: start;
-  font-size: 62.5%;
-  height: 33rem;
-  margin: 2rem;
+    display: -webkit-box;
+    width: 85rem;
+    grid-gap: 6rem;
+    grid-template-columns: repeat(auto-fit, minmax(30rem, 1fr));
+    align-items: start;
+    font-size: 62.5%;
+    height: 33rem;
+    margin: 2rem;
 }
+
 @media (max-width: 60em) {
-  .grid {
-    grid-gap: 3rem;
-  }
+    .grid {
+        grid-gap: 3rem;
+    }
 }
+
 .grid__item {
-  background-color: #fff;
-  border-radius: 0.4rem;
-  overflow: hidden;
-  box-shadow: 0 3rem 6rem rgba(0, 0, 0, 0.1);
-  cursor: pointer;
-  transition: 0.2s;
+    background-color: #fff;
+    border-radius: 0.4rem;
+    overflow: hidden;
+    box-shadow: 0 3rem 6rem rgba(0, 0, 0, 0.1);
+    cursor: pointer;
+    transition: 0.2s;
+    width: 37%;
+    height: 152%;
 }
+
 .grid__item:hover {
-  transform: translateY(-0.5%);
-  box-shadow: 0 4rem 8rem rgba(0, 0, 0, 0.2);
+    transform: translateY(-0.5%);
+    box-shadow: 0 4rem 8rem rgba(0, 0, 0, 0.2);
 }
 
 .card__img {
-  display: block;
-  width: 100%;
-  height: 18rem;
-  object-fit: cover;
+    display: block;
+    width: 100%;
+    height: 18rem;
+    object-fit: cover;
 }
+
 .card__content {
-  padding: 2rem 4rem;
+    padding: 2rem 4rem;
 }
+
 .card__header {
-  font-size: 3rem;
-  font-weight: 500;
-  color: #0d0d0d;
-  margin-bottom: 1.5rem;
+    font-size: 3rem;
+    font-weight: 500;
+    color: #0d0d0d;
+    margin-bottom: 1.5rem;
 }
+
 .card__text {
-  font-size: 1.5rem;
-  letter-spacing: 0.1rem;
-  line-height: 1.7;
-  color: #3d3d3d;
-  margin-bottom: 2.5rem;
+    font-size: 1.5rem;
+    letter-spacing: 0.1rem;
+    line-height: 1.7;
+    color: #3d3d3d;
+    margin-bottom: 2.5rem;
 }
+
 .card__btn {
-  display: block;
-  width: 100%;
-  padding: 1.5rem;
-  font-size: 2rem;
-  text-align: center;
-  color: #3363ff;
-  background-color: #e6ecff;
-  border: none;
-  border-radius: 0.4rem;
-  transition: 0.2s;
-  cursor: pointer;
-  margin: 1.5rem;
+    display: block;
+    width: 100%;
+    padding: 1.5rem;
+    font-size: 2rem;
+    text-align: center;
+    color: #3363ff;
+    background-color: #e6ecff;
+    border: none;
+    border-radius: 0.4rem;
+    transition: 0.2s;
+    cursor: pointer;
+    margin: 1.5rem;
 }
+
 .card__btn span {
-  margin-left: 1rem;
-  transition: 0.2s;
+    margin-left: 1rem;
+    transition: 0.2s;
 }
+
 .card__btn:hover, .card__btn:active {
-  background-color: #dce4ff;
+    background-color: #dce4ff;
 }
+
 .card__btn:hover span, .card__btn:active span {
-  margin-left: 1.5rem;
+    margin-left: 1.5rem;
 }
