+++ templates/edit_expense.html
@@ -16,7 +16,7 @@
     <div class="add_group">
         <div class="row">
             <div class="col-md-12">
-                <form action="/split/process_edit/" method="POST">
+                <form action="/split/edit_expense/" method="POST">
                     {% csrf_token %}
                     <fieldset>
                         <legend><span class="number"></span> Add Expense</legend>
