+++ userprofile/urls.py
@@ -1,7 +1,8 @@
+from django.contrib.auth.decorators import login_required
 from django.urls import path
 
 from userprofile.views import UserView
 
 urlpatterns = [
-    path('user_profile/', UserView.as_view(), name='user_profile'),
+    path('user_profile/', login_required(UserView.as_view(), login_url='/'), name='user_profile'),
 ]
