+++ userprofile/views.py
@@ -1,13 +1,12 @@
-from django.contrib.auth.decorators import login_required
+from django.contrib.auth import authenticate, login
 from django.contrib.auth.models import User
+from django.core.paginator import Paginator
 from django.shortcuts import render
-from django.utils.decorators import method_decorator
 from django.views import View
 from group.models import Group
 from userprofile.models import Profile
 
 
-@method_decorator(login_required(login_url='/login/'), name='dispatch')
 class UserView(View):
     template_name = 'userprofile.html'
 
@@ -65,6 +64,7 @@ class UserView(View):
         else:
             profile = Profile.objects.get(user=request.user)
             user = User.objects.get(id=request.user.id)
+            username = user.username
             if picture:
                 profile.avatar = picture
             user.first_name = first_name
@@ -74,8 +74,11 @@ class UserView(View):
             user.set_password(new_password)
             profile.save()
             user.save()
-
+            new_user = authenticate(request, username=username, password=new_password)
+            login(request, new_user)
         groups = Group.objects.filter(users=request.user.id)
-        profile = Profile.objects.filter(user=request.user).first()
-        context = {'groups': groups, 'profile': profile}
+        paginator = Paginator(groups, 6)  # Create a Paginator instance with 6 items per page
+        page_number = request.GET.get('page', 1)  # Get the current page number from the request's query parameters
+        page_obj = paginator.get_page(page_number)
+        context = {'page_obj': page_obj}
         return render(request, 'dashboard.html', context)
